requirements.txt
SQLAlchemy
pydantic
python-dotenv
psycopg2